requirements.txt
Django==1.8.4
django-filter==0.11.0
djangorestframework==3.2.3
pytz==2015.4
wsgiref==0.1.2
